- Use correct type (bug #32720)
git-svn-id: http://svn.php.net/repository/php/php-src/php/php-src/trunk@184418 c90b9560-bf6c-de11-be94-00142212c4b1

+++ ext/snmp/snmp.c
@@ -351,9 +351,9 @@ static void php_snmp_internal(INTERNAL_FUNCTION_PARAMETERS, int st,
 	struct snmp_pdu *pdu=NULL, *response;
 	struct variable_list *vars;
 	oid name[MAX_NAME_LEN];
-	int name_length;
+	size_t name_length;
 	oid root[MAX_NAME_LEN];
-	int rootlen = 0;
+	size_t rootlen = 0;
 	int gotroot = 0;
 	int status, count;
 	char buf[2048];
